utils/mqtt-sub/requirements.txt
bitarray==3.8.1
bitstring==4.4.0
cffi==2.0.0
click==8.3.2
construct==2.10.70
cryptography==46.0.7
esp_idf_nvs_partition_gen==0.2.0
esp_secure_cert_tool==2.6.0
esptool==5.2.0
intelhex==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
paho-mqtt==2.1.0
pycparser==3.0
Pygments==2.20.0
pyserial==3.5
PyYAML==6.0.3
reedsolo==1.7.0
rich==15.0.0
rich-click==1.9.7
tibs==0.5.7
